+++ plugin/CustomShop/src/pch.h
@@ -6,6 +6,13 @@
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
 
+#ifdef min
+#undef min
+#endif
+#ifdef max
+#undef max
+#endif
+
 #include <string>
 #include <vector>
 #include <memory>